backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-CORS==4.0.1
flask-bcrypt==1.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
google-auth==2.29.0
requests==2.31.0
email-validator==2.1.1
